pyproject.toml
[project]
name = "leave-application-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "azure-cosmos>=4.14.5",
  "azure-functions>=1.24.0",
  "azure-identity>=1.25.1",
  "docx2txt>=0.9",
  "fastapi[standard]>=0.128.0",
  "fastmcp>=2.14.2",
  "langchain>=1.2.4",
  "langchain-community>=0.4.1",
  "langchain-core>=1.2.7",
  "langchain-openai>=1.1.7",
  "langgraph>=1.0.5",
  "pydantic>=2.12.5",
  "pypdf>=6.6.2",
  "python-docx>=1.2.0",
  "requests>=2.32.5",
  "sentence-transformers>=5.2.2",
  "torch>=2.10.0",
]
